Machine Learning Assignment-4
Support Vector Machines
Assignment-4 Solutions
Fraud in Wine
A distribution company that was recently a victim of fraud has completed an audit of various samples of wine through the use of chemical analysis on samples.
The distribution company specializes in exporting extremely high quality, expensive wines, but was defrauded by a supplier who was attempting to pass off cheap,
low-quality wine as higher grade wine. The distribution company has hired you to attempt to create a machine learning model that can help detect low-quality
(a.k.a "fraud") wine samples. They want to know if it is even possible to detect such a difference.¶